Major upgrade to Pi Fighter Server
Added in fighters and their attacks.  It can use attacks recorded in other fights as the opponent attacks in another fight.  This allows players to fight previous players or themselves in a different session.  Also added some rudimentary regeneration of health and increase in the player's base health when they win a match.